ansible/make-ca-cert: Clean up verbosity, error on no-public-ip
In a configuration I was testing in OpenStack, there was no
public floating IP attached to the master.  This caused `make-ca-cert.sh`
to fail in a very obscure way while doing a `cp` of a nonexistent
cert, with `stderr` redirected to `/dev/null`.

Now apparently there's a reason for directing `easyrsa` `stderr`
to `/dev/null` - OpenSSL is just verbose.

But for the other things, e.g. curl has a `-s` flag.  And AFAICS there
was never any reason to hide errors from `cp`.

Anyways, we fix the real problem here by cleanly erroring out if we
fail to get the public IP from any of the cloud providers, before we
even try to make a cert.